Вносит правки на замечания после консультации
